% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rental.R
\name{get_listings}
\alias{get_listings}
\title{Access to a hypothetical rental listings database}
\usage{
get_listings(start_date, end_date, region, beds = NA, size = NA,
  sanity = c(400, 10000), filter = "all", quiet = TRUE)
}
\arguments{
\item{start_date}{starting date for listings (inclusive)}

\item{end_date}{ending date for listings (exclusive)}

\item{region}{a geographic region}

\item{beds}{optional, filter by number of bedrooms. Can be integer or vector with several values}

\item{size}{optional, filter by size, vector with min and max size}

\item{sanity}{optional, sanity filter to exclude extremely low or high price listings}

\item{filter}{optional, possible values are "all", "furnished" or "unfurnished".}

\item{quiet}{optional, displays some debug info if true}
}
\description{
The function expect that the following option variables are set
rental.dbname, rental.host, rental.user, rental.password
in order to access the remote postgres database
Set thesee using e.g. options(renta.dbname="xxxxx") in your .Rprofile
}
\examples{
get_listings(start_date="2017-08-01", end_date="2017-09-01", region=geometry, filter = 'unfurnished')
}
